# perf: optimize hydrateRuntimeState