--- a/.pnp.js
+++ b/.pnp.js
@@ -41674,7 +41674,7 @@ function hydrateRuntimeState(data, {
         discardFromLookup: packageInformationData.discardFromLookup || false,
         _packageLocation: undefined,
 
-        // we only need this computed for packages that are used by the running script
+        // we only need this for packages that are used by the currently running script
         // this is a lazy getter because `ppath.join` has some overhead
         get packageLocation() {
           // We use ppath.join instead of ppath.resolve because: