**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

The serialized state map was being iterated twice. Also `ppath.join` seems to have a bit of overhead.

While testing on the `berry` repo, the changes in this PR shave off about 20ms in boot-up time for the pnp hook:

Before the change:
![image](https://user-images.githubusercontent.com/697707/99386702-8e94d700-28db-11eb-9431-fc384246f2c6.png)

After the change:
![image](https://user-images.githubusercontent.com/697707/99386729-9785a880-28db-11eb-8a80-acc09962caec.png)


**How did you fix it?**
<!-- A detailed description of your implementation. -->

Moved the logic from the second loop into the first one.
Inlined the `ppath.join` call, based on assumptions that the path is relative.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically tested. -->
<!--- This checklist exists to help you remember doing the chores when you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).

<!-- See https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check -i` -->
- [x] I have set the packages that need to be released for my changes to be effective.

<!-- The "Testing chores" workflow validates that your PR follows our guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets reviewed.