@andreialecu 2020-11-19
@arcanis I reverted the `fastJoinPath` and kept only the loop optimization.

I thought about it a bit more and I think there's a better solution for the `ppath.join()` overhead, should that ever become a problem:

* Instead of making paths absolute when hydrating the state, `packageLocation` could be stored relative in the state, as is.
* Functions that need to work with an absolute `packageLocation` can then call `ppath.join(runtimeState.basePath, packageLocation)` themselves.
* Since in big mono repos it is likely that only a subset of paths need to be resolved, and only once (for `require`), this should be pretty fast.

There are only a few such calls to fix in `berry`, but it would be a breaking change if anyone uses `@yarnpkg/pnp` directly.