@merceyz 2020-11-19
Not saying to do it but instead of refactring a lot of stuff you could make it a getter to compute it lazily, don't know if that would improve much as I haven't benchmarked it