@andreialecu 2020-11-19
@merceyz I benchmarked your suggestion and it does improve execution speed of `hydrateRuntimeState` by about 10ms for `berry` itself. But... the resulting implementation is pretty ugly 🙂

```ts
      const packageInformation = {
        packageDependencies: new Map(packageInformationData.packageDependencies),
        packagePeers: new Set(packageInformationData.packagePeers),
        linkType: packageInformationData.linkType,
        discardFromLookup: packageInformationData.discardFromLookup || false
      };
      Object.defineProperty(packageInformation, 'packageLocation', {
        enumerable: true,
        get: ppath.join.bind(null, absolutePortablePath, packageInformationData.packageLocation),
      });
      return [packageReference, packageInformation];
```

**With** getter:
![image](https://user-images.githubusercontent.com/697707/99665601-88852e80-2a72-11eb-9e62-c06b0fcff731.png)

Without:
![image](https://user-images.githubusercontent.com/697707/99665763-bd918100-2a72-11eb-8cb2-aee8c929764f.png)