@merceyz 2020-11-19
You'll need to change your benchmark to take into account when the getter is executed

> But... the resulting implementation is pretty ugly

Doesn't have to be
```diff
diff --git a/packages/yarnpkg-pnp/sources/loader/hydrateRuntimeState.ts b/packages/yarnpkg-pnp/sources/loader/hydrateRuntimeState.ts
index c13f7184..e975ae60 100644
--- a/packages/yarnpkg-pnp/sources/loader/hydrateRuntimeState.ts
+++ b/packages/yarnpkg-pnp/sources/loader/hydrateRuntimeState.ts
@@ -20,7 +20,9 @@ export function hydrateRuntimeState(data: SerializedState, {basePath}: HydrateRu
         // We use ppath.join instead of ppath.resolve because:
         // 1) packageInformationData.packageLocation is a relative path when part of the SerializedState
         // 2) ppath.join preserves trailing slashes
-        packageLocation: ppath.join(absolutePortablePath, packageInformationData.packageLocation),
+        get packageLocation() {
+          return ppath.join(absolutePortablePath, packageInformationData.packageLocation);
+        },
         packageDependencies: new Map(packageInformationData.packageDependencies),
         packagePeers: new Set(packageInformationData.packagePeers),
         linkType: packageInformationData.linkType,

```