@andreialecu 2020-11-19
TIL that you can use getters on plain objects 😮 

Alright, here's a benchmark that loads a bunch of stuff:

```js
//index.js

require(`@babel/core`);

require(`babel-jest`);
require(`clipanion`);
require(`eslint`);
require(`eslint-plugin-arca`);
require(`eslint-plugin-react`);
require(`github-api`);
require(`jest`);
require(`jest-environment-node`);
require(`jest-junit`);
require(`micromatch`);
require(`tslib`);
require(`typescript`);
```

Without the getter:
```
hyperfine --warmup 2 'node -r ./.pnp.js index.js'
Benchmark #1: node -r ./.pnp.js index.js
  Time (mean ± σ):     896.7 ms ±  25.4 ms    [User: 1.181 s, System: 0.092 s]
  Range (min … max):   876.2 ms … 954.6 ms    10 runs
```

With getter:
```
hyperfine --warmup 2 'node -r ./.pnp-getter.js index.js' 
Benchmark #1: node -r ./.pnp-getter.js index.js
  Time (mean ± σ):     874.2 ms ±   7.4 ms    [User: 1.169 s, System: 0.069 s]
  Range (min … max):   858.4 ms … 883.5 ms    10 runs
```

Current master (without this PR):
```
hyperfine --warmup 2 'node -r ./.pnp.js index.js'
Benchmark #1: node -r ./.pnp.js index.js
  Time (mean ± σ):     914.7 ms ±  29.6 ms    [User: 1.208 s, System: 0.091 s]
  Range (min … max):   880.9 ms … 991.6 ms    10 runs
```

They all vary a bit, but it seems that the implementation that uses the `getter` always comes up a little ahead. It's surprising because I would have expected that allocating the getter function would have some overhead, more so than a simple `ppath.join` call.

There's no relevant memory usage difference (see `Maximum resident set size (kbytes)`):

```
$ /usr/bin/time -v node -r ./.pnp.js index.js         
        Command being timed: "node -r ./.pnp.js index.js"
        User time (seconds): 1.27
        System time (seconds): 0.05
        Percent of CPU this job got: 140%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:00.94
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 179816
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 31806
        Voluntary context switches: 1227
        Involuntary context switches: 7
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
$ /usr/bin/time -v node -r ./.pnp-getter.js index.js
        Command being timed: "node -r ./.pnp-getter.js index.js"
        User time (seconds): 1.15
        System time (seconds): 0.05
        Percent of CPU this job got: 140%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:00.85
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 178768
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 31631
        Voluntary context switches: 1172
        Involuntary context switches: 4
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

I have pushed a commit with the getter, hopefully it's fine. Otherwise feel free to remove it, `Allow edits from maintainers` is on, as always.